# openapi.yaml
openapi: 3.1.0
info:
  title: Flyleaf API
  description: The social reading tracker backend (PRD §24, Architecture §6).
  version: 0.0.1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access JWT (FN-63)
...
paths:
...
  /admin/auth/login:
    post:
      summary: Admin login with password and mandatory 2FA
      tags:
        - Admin Auth
      description: Authenticates administrator or moderator credentials with TOTP 2FA,
        issuing a scoped Admin JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 1
                totp_code:
                  type: string
                  minLength: 6
                  maxLength: 10
                  description: 6-digit TOTP code or backup code
              required:
                - email
                - password
                - totp_code
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Admin JWT scoped to flyleaf-admin
                  admin:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                        format: email
                      role:
                        type: string
                        enum:
                          - admin
                          - moderator
                    required:
                      - id
                      - email
                      - role
                required:
                  - token
                  - admin
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
  /admin/auth/me:
    get:
      summary: Get current admin session
      tags:
        - Admin Auth
      description: Returns the authenticated administrator or moderator viewer context.
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                        format: email
                      role:
                        type: string
                        enum:
                          - admin
                          - moderator
                    required:
                      - id
                      - email
                      - role
                required:
                  - admin
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
  /admin/auth/2fa/setup:
    post:
      summary: Initialize 2FA setup for admin account
      tags:
        - Admin Auth
      description: Generates a new TOTP secret, otpauth URI, and one-time backup codes.
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                  otpauth_uri:
                    type: string
                  backup_codes:
                    type: array
                    items:
                      type: string
                required:
                  - secret
                  - otpauth_uri
                  - backup_codes
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
  /admin/auth/2fa/verify:
    post:
      summary: Verify and activate 2FA setup
      tags:
        - Admin Auth
      description: Verifies the first 6-digit TOTP code to confirm two-factor setup.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
              required:
                - code
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
...
servers:
  - url: /v1
    description: Default API path
...